Create more integration tests Fixes #1
